# olympapi-validation-suite/docs/COLLECTIONS/OlympAPI Test Server- swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OlympAPI Test Server",
        "version": "1.0.0",
        "description": "A dedicated test server covering all OlympAPI features: HTTP methods, auth, body types, chaining, scripts, SSL, and proxy."
    },
    "servers": [
        {
            "url": "http://localhost:3000",
            "description": "HTTP"
        },
        {
            "url": "https://localhost:3443",
            "description": "HTTPS (self-signed)"
        }
    ],
    "components": {
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "token",
                "description": "Use \"test-token-123\" for /auth/bearer, or the token from POST /auth/login for /auth/me and /users/:id"
            },
            "BasicAuth": {
                "type": "http",
                "scheme": "basic",
                "description": "Use admin:password123"
            },
            "ApiKeyHeader": {
                "type": "apiKey",
                "in": "header",
                "name": "X-API-Key",
                "description": "Use \"api-key-xyz\""
            },
            "ApiKeyQuery": {
                "type": "apiKey",
                "in": "query",
                "name": "api_key",
                "description": "Use \"api-key-xyz\""
            }
        }
    },
...
    "paths": {
...
        "/cookies/set": {
            "get": {
                "summary": "Set one or more test cookies from query parameters",
                "tags": [
                    "Cookies"
                ],
                "parameters": [
                    {
                        "in": "query",
                        "name": "session",
                        "schema": {
                            "type": "string",
                            "example": "abc123"
                        },
                        "description": "Session cookie value to set. Defaults to abc123 when no cookie query is provided."
                    },
                    {
                        "in": "query",
                        "name": "maxAge",
                        "schema": {
                            "type": "integer",
                            "example": 3600
                        },
                        "description": "Optional Max-Age attribute in seconds."
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cookies were set through Set-Cookie headers"
                    }
                }
            }
        },
...
        "/data/user/{id}": {
            "get": {
                "summary": "Get a single user by ID",
                "tags": [
                    "Data"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User found"
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            }
        },
        "/users/{id}": {
            "get": {
                "summary": "Get user by ID (chaining target — requires Bearer token from /auth/login)",
                "tags": [
                    "Data"
                ],
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User profile"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            }
        },
...
        "/headers/required": {
            "get": {
                "summary": "Require X-Custom-Header header",
                "tags": [
                    "Headers"
                ],
                "parameters": [
                    {
                        "in": "header",
                        "name": "X-Custom-Header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Header received"
                    },
                    "400": {
                        "description": "Missing X-Custom-Header"
                    }
                }
            }
        },
        "/params": {
            "get": {
                "summary": "Echo query parameters",
                "tags": [
                    "Params"
                ],
                "parameters": [
                    {
                        "in": "query",
                        "name": "name",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "in": "query",
                        "name": "page",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Received query params"
                    }
                }
            }
        },
        "/params/required": {
            "get": {
                "summary": "Require name and page query parameters",
                "tags": [
                    "Params"
                ],
                "parameters": [
                    {
                        "in": "query",
                        "name": "name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "in": "query",
                        "name": "page",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Valid params received"
                    },
                    "400": {
                        "description": "Missing required params"
                    }
                }
            }
        },
...
        "/status/{code}": {
            "get": {
                "summary": "Respond with the given HTTP status code",
                "tags": [
                    "Status"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "code",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 200,
                            "maximum": 599
                        },
                        "description": "HTTP status code to return"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns the requested status code (actual status varies)"
                    }
                }
            }
        },
        "/delay/{ms}": {
            "get": {
                "summary": "Respond after a delay",
                "tags": [
                    "Status"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "ms",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 0,
                            "maximum": 3000
                        },
                        "description": "Delay in milliseconds (max 3000)"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Delayed response"
                    }
                }
            }
        }
    }
}